requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.1.1
cycler==0.12.0
et-xmlfile==1.1.0
filelock==3.12.4
fonttools==4.43.0
fsspec==2023.9.2
idna==3.4
importlib-resources==6.1.0
Jinja2==3.1.2

kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4

natsort==8.4.0
opencv-python==4.8.1.78
openpyxl==3.1.2
pandas==2.0.3
Pillow==9.3.0
protobuf==4.24.4
pyparsing==3.1.1
pytz==2023.3.post1
requests==2.28.1
sympy==1.12
tensorboardX==2.6.2.2
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.66.1

